+++ client/src/primitive/point/point.rs
@@ -3,8 +3,8 @@ use winit::dpi::PhysicalPosition;
 
 #[derive(Clone, Copy, Debug, PartialEq)]
 pub struct Point {
-    pub x: i32,
-    pub y: i32,
+    pub x: f32,
+    pub y: f32,
 }
 
 impl Point {
@@ -20,32 +20,29 @@ impl Point {
 
     pub fn rotate(&self, angle: f32) -> Point {
         // Приведем точку к началу координат
-        let x = self.x as f32;
-        let y = self.y as f32;
+        let x = self.x;
+        let y = self.y;
         let x2 = x * angle.cos() - y * angle.sin();
         let y2 = x * angle.sin() + y * angle.cos();
-        return Point {
-            x: x2 as i32,
-            y: y2 as i32,
-        };
+        return Point { x: x2, y: y2 };
     }
 
     pub fn is_zero(&self) -> bool {
-        self.x == 0 && self.y == 0
+        self.x == 0.0 && self.y == 0.0
     }
 }
 
 impl Default for Point {
     fn default() -> Self {
-        Point { x: 0, y: 0 }
+        Point { x: 0.0, y: 0.0 }
     }
 }
 
 impl From<PhysicalPosition<f64>> for Point {
     fn from(position: PhysicalPosition<f64>) -> Self {
         Self {
-            x: position.x as i32,
-            y: position.y as i32,
+            x: position.x as f32,
+            y: position.y as f32,
         }
     }
 }
@@ -77,23 +74,23 @@ mod tests {
     use super::*;
     #[test]
     fn test_add() {
-        let p1 = Point { x: 1, y: 1 };
-        let p2 = Point { x: 1, y: 1 };
-        assert_eq!(p1 + p2, Point { x: 2, y: 2 })
+        let p1 = Point { x: 1.0, y: 1.0 };
+        let p2 = Point { x: 1.0, y: 1.0 };
+        assert_eq!(p1 + p2, Point { x: 2.0, y: 2.0 })
     }
 
     #[test]
     fn test_substract() {
-        let p1 = Point { x: 2, y: 2 };
-        let p2 = Point { x: 1, y: 1 };
-        assert_eq!(p1 - p2, Point { x: 1, y: 1 })
+        let p1 = Point { x: 2.0, y: 2.0 };
+        let p2 = Point { x: 1.0, y: 1.0 };
+        assert_eq!(p1 - p2, Point { x: 1.0, y: 1.0 })
     }
 
     #[test]
     fn is_zero() {
-        let p1 = Point { x: 0, y: 0 };
-        let p2 = Point { x: 0, y: 1 };
-        let p3 = Point { x: 1, y: 1 };
+        let p1 = Point { x: 0.0, y: 0.0 };
+        let p2 = Point { x: 0.0, y: 1.0 };
+        let p3 = Point { x: 1.0, y: 1.0 };
 
         assert_eq!(p1.is_zero(), true);
         assert_eq!(p2.is_zero(), false);
